Extract CalendarCell into separate component
